Created signup page
I created signup page. I did not get into CSS of the button at the bottom quite yet as I was adding ScrollView to work with it, but I will look into it after.
I got rid of the header that leads to account on the page as there is no my account yet, also added the feature to password input so it's secure to enter it.